+++ crates/bonsaidb-client/src/client/wasm_websocket_worker.rs
@@ -39,13 +39,13 @@ async fn create_websocket<
     subscribers: SubscriberMap,
 ) {
     subscribers.clear().await;
-    
+
     // Receive the next/initial request when we are reconnecting.
     let initial_request = match request_receiver.recv_async().await {
         Ok(r) => r,
         Err(_) => return,
     };
-    
+
     // In wasm we're not going to have a real loop. We're going create a
     // websocket and store it in JS. This will allow us to get around Send/Sync
     // issues since each access of the websocket can pull it from js.