+++ crates/bonsaidb/examples/server.rs
@@ -45,7 +45,9 @@ async fn main() -> anyhow::Result<()> {
     {
         let server = server.clone();
         tokio::spawn(async move {
-            server.listen_for_http_on("localhost:8080").await
+            server
+                .listen_for_websockets_on("localhost:8080", false)
+                .await
         });
     }